// src/core/exceptions/constants/verification-error.enum.ts
export enum VerificationError {
    CAN_NOT_APPROVE = "can-not-approve",
    CAN_NOT_DENIED = "can-not-denied",
    CAN_NOT_GET_VERIFICATION_LIST = "can-not-get-verification-list",
    CAN_NOT_GET_VERIFICATION_DETAIL = "can-not-get-verification-detail",
    CAN_NOT_UPDATE_VERIFICATION_DETAIL = "can-not-update-verification-detail"
}